ide: implements projects
File open, save, etc menu actions for projects.
Adds project editing preference widget.
No longer saves "standalone" option in ide preferences, instead display the 'excludeDefaultPaths' option in 'interpreter' pane of the preferences dialog.
ide can be in two different modes, normal or project, this is decided by the '"IDE/interpreter/project' boolean key in the ide preferences. This key is set to true when opening a project and set to false and closing the project.